implement confirm btn for Phonebook step9 (2.17)

+++ part2/ex2.6/src/App.js
@@ -9,11 +9,13 @@ const Entries = ({persons, personsSet}) => {
 
 const Entry = ({person, persons, personsSet}) => {
   const removeEntry = (entry) => {
+    if (window.confirm(`Delete ${entry.name}?`)) {
     entryService
       .deleteEntry(entry)
       .then(response => {
         personsSet(persons.filter(p => p.id !== person.id))
       })
+    }
   }
 
   return (
